test(analytics): move sysadmin tests to analytics/community
Phase G commit 9。__tests__/unit/sysadmin/ → __tests__/unit/analytics/community/
にテストファイルを物理移動 (`git mv` でリネーム検出を効かせる)。

  __tests__/unit/sysadmin/  →  __tests__/unit/analytics/community/
    aggregations.test.ts
    classifiers.test.ts
    pagination.test.ts
    presenter.test.ts
    service.test.ts
    util.test.ts
    memberListCursor.test.ts
    fixtures.ts

各テストの `@/__tests__/unit/sysadmin/fixtures` import を新ロケーションに
追従更新。テスト本体のロジックは未変更 (commit 2.5 / 3 / 7 で
analytics/community/* に対応する import に切替済み)。

Phase G の最後の作業。これで sysadmin → analytics 解体 + report の
admin* 命名規則修正が完了:
- 旧 sysadmin/ ドメインは controller/resolver.ts と schema/ のみ残り、
  新 analytics への薄い shim として動作。
- report/ の admin* query は新名前 (reportsAll / reportSummaries /
  reportTemplateFeedbacks / reportTemplateFeedbackStats) で並行実装、
  旧名前は @deprecated。
- 91 unit tests green、typecheck pass、動作変更ゼロ。

Phase H (別 PR、本 PR 完了後): frontend が新 query に切替後、
旧 sysadmin/ + 旧 admin* query / type を削除。
