+++ src/__tests__/unit/analytics/community/aggregations.test.ts
@@ -9,7 +9,7 @@ import {
 } from "@/application/domain/analytics/community/aggregations";
 import { DEFAULT_SEGMENT_THRESHOLDS } from "@/application/domain/analytics/community/classifiers";
 import type { AnalyticsMonthlyActivityRow } from "@/application/domain/analytics/community/data/type";
-import { member } from "@/__tests__/unit/sysadmin/fixtures";
+import { member } from "@/__tests__/unit/analytics/community/fixtures";
 
 // ========================================================================
 // computeStageCounts: cumulative (tier2 includes tier1)
